[3기-C] 서예원 SpringBoot Part2 Weekly Mission(1차) 제출합니다

+++ src/main/java/org.prgrms/console/Validator.java
@@ -2,9 +2,9 @@ package org.prgrms.console;
 
 public class Validator {
 
-  public static long parseNum(String input) {
+  public static int parseNum(String input) {
     try {
-      return Long.parseLong(input);
+      return Integer.parseInt(input);
     } catch (NumberFormatException e) {
       throw new RuntimeException("Only numbers can be entered. *current input value: " + input);
     }
